Correct rule for associated type name conflicts
The API Guidelines document specifies that, when both an associated type and the protocol it's constrained to ought to have the same name, you should resolve the conflict by adding a -`Type` suffix to the associated type. This was the rule in Swift 2, but ever since SE-0023 was approved, we've been following a different pattern—adding a -`Protocol` suffix to the protocol. It appears the guidelines were never updated to reflect this change.